Move jobs into subfolders
