+++ src/brain/slack/triggerPubSub/index.test.ts
@@ -1,7 +1,7 @@
 import { createSlackAppMention } from '@test/utils/createSlackAppMention';
 
 import { buildServer } from '@/buildServer';
-import * as slackScoresFunctions from '@/jobs/slackScores';
+import * as slackScoresFunctions from '@/jobs/slackScores/slackScores';
 
 import { TEAM_OSPO_CHANNEL_ID } from '../../../config';
