+++ src/jobs/gocdPausedPipeline/gocdPausedPipelineBot.test.ts
@@ -4,8 +4,8 @@ import moment from 'moment-timezone';
 import { GETSENTRY_ORG } from '@/config';
 import { GoCDDashboardResponse } from '@/types/gocd';
 
-import * as gocdAPI from '../api/gocd/index';
-import { bolt } from '../api/slack/__mocks__';
+import * as gocdAPI from '../../api/gocd/index';
+import { bolt } from '../../api/slack/__mocks__';
 
 import { triggerPausedPipelineBot } from './gocdPausedPipelineBot';
